==> code/summarizing_data/01_filtering_summarizing_camera_data.Rmd <==
---
title: "Filtering and summarizing camera data for analysis"
output: html_document
knit: (function(inputFile, encoding) {
      out_dir <- "knitted_markdown_files";
      rmarkdown::render(inputFile,
                        encoding=encoding,
                        output_dir=file.path(dirname(inputFile), out_dir))})
---

**Goal**

Filter camera data to remove:

- Videos where no flowers were present
- Videos from afternoon of covering day; these likely reflect behavioral adjustment by hummingbirds, not the experimental outcome or 'normal' visitation

Then, subset this dataset ('full dataset') into different datasets for different analyses.

1. Experimental dataset

**Purpose:** Investigate hummingbird responses to experimental Heliconia removal, following Before-After-Control-Impact design

- Needs cameras to have data from pre AND post periods, because I'm interested in the pre-to-post change
- Only includes data from pre and post periods, not videos reviewed outside of those periods

2. 'Normal' (unmanipulated) dataset

**Purpose:** Understand 'normal' visitation patterns in this study system. Use for camera meta-network, which summarizes visitation across all sites and years.

- Not bound by the requirements in #1. 
- Does not include data from the 'post' period of treatment replicates; do not want data affected by *Heliconia* removal

3. Sampling method dataset

**Purpose:** Analyze how sampling method (pollen vs. camera) affects specialization, using parallel networks collected during the same time and at the same site. Also examine the extent of correlation between pollen and camera approaches.

- Only includes data from 'pre' periods, because want data sampled from the dates ('post' period is not an option because it's affected by *Heliconia* removal)

```{r setup, include=FALSE}
knitr::opts_chunk$set(message = FALSE, warning = FALSE)

library(tidyr)
library(dplyr)
library(lubridate)
library(ggplot2)
```

## Step 1: Import camera data + prepare for filtering
```{r step1}
## 1. Create column for unique camera ID
## 2. Convert start/end columns into period objects, so can treat them as times when filtering out afternoon videos from covering days
data <- read.csv("../../data/import/data/cameras/Camera_data_with_flower_counts_2016-2018.csv") %>%
  mutate(camera_id = paste(year, patch, plant_species, camera_num, sep = "_")) %>% 
  mutate(across(c(video_start, video_end), ~lubridate::hm(.))) %>%
  mutate(across(c(sighting_start, sighting_end), ~lubridate::hms(.)))
```

## Step 2: Run some initial checks before filtering 
```{r step2}
## Check strings that failed to parse in previous step
check01 <- data %>%
  filter(if_any(c(video_start, video_end), ~is.na(.)))

## Are there any missing start/end times for sightings within a video?
## Filter to rows with sighting start/end times (no sightings = NA = no times to parse = warning message)
check02 <- data %>%
  filter(if_any(c(sighting_start, sighting_end), ~is.na(.))) %>%
  filter(sightings_yes_or_no == "Y")

## Are there videos from non-priority dates?
## "Priority" = days in between capture sessions, omitting the day after cover day (for treatment sites) or its equivalent (for control sites)
check03 <- data %>%
  filter(video_priority != 1) %>%
  select(contains("day"), everything())

## Are there videos without any visible flowers?
check06 <- data %>%
  filter(flowers_camera_video == 0)

## It doesn't really make sense to analyze videos with no visible flowers, but out of curiosity...do hummingbirds ever visit anyway?
## "Sighting" = hummingbird appears on camera, "Visit" = hummingbird appears to drink from flower
check04 <- data %>%
  filter(flowers_camera_video == 0 & sightings_yes_or_no == "Y" & visit_type != "none") 

## Are there sightings from the afternoon of cover day? 
## Covering is a big disturbance to the site, so we should NOT use these data...and especially should not consider them 'pre' data!
check05 <- data %>%
  filter(treatment_cover_day == 1) %>%
  filter(video_end >= lubridate::hm("11:30")) %>%
  filter(sightings_yes_or_no == "Y")

## Which species have NA for flowers?
check06 <- data %>%
  filter(is.na(flowers_camera_video)) %>%
  group_by(plant_species) %>%
  summarise(n = n()) %>%
  ungroup() %>%
  arrange(desc(n))
```

## Step 3: Filter data for all analyses

Remove videos with no flowers on camera (videos were typically not reviewed if they had no flowers, but some were reviewed anyway).
```{r step3a}
## I will still include NAs, since some plants never had flowers visible on video (i.e., MARA-VER, GINGY)
data02 <- data %>%
  filter(flowers_camera_video > 0 | is.na(flowers_camera_video))
```    

Remove sightings from afternoon of cover day (treatment sites only). Here, 'afternoon' = videos starting after 11:30 AM
```{r step3b}
## Videos that start (and end) after 11:30 (will just remove these)
treatment_videos_pm <- data02 %>%
  filter(treatment_cover_day == 1 & control_treatment == "treatment") %>%
  filter(video_start > hm("11:30"))

## Videos that span the cutoff time (will change end time to 11:30 and remove sightings after this time)
treatment_videos_am_pm <- data02 %>%
  filter(treatment_cover_day == 1 & control_treatment == "treatment") %>%
  filter(video_start < hm("11:30")) %>%
  filter(video_end > hm("11:30"))

## Sightings from afternoon of covering
treatment_sightings_pm <- treatment_videos_am_pm %>%
  filter(sighting_end > hm("11:30"))

## 1. Remove videos from afternoon of cover day (i.e., videos starting after 11:30AM)
## 2. Adjust end time of videos from afternoon of cover day
## 3. Recalculate video length with updated end times
## 4. Make 'adjusted' video length the new video length
data03 <- data02 %>% 
  filter(!(file_id %in% treatment_videos_pm$file_id)) %>% 
  filter(!(sighting_id %in% treatment_sightings_pm$sighting_id)) %>% 
  mutate(video_end_adjust = dplyr::if_else(file_id %in% treatment_videos_am_pm$file_id, hm("11:30"), video_end)) %>%
  mutate(video_length_adjust = as.numeric(video_end_adjust - video_start, "hours")) %>%
  select(row_id:video_length, video_end_adjust, video_length_adjust, everything()) %>%
  select(-video_length) %>%
  rename(video_length = video_length_adjust)
```

Calculate total number of video hours in full dataset (from which smaller datasets are created)
```{r step3c}
hours_per_camera_date <- data03 %>%
  distinct(year, patch, control_treatment, camera_num, date_video, exp_phase, file_id, video_length) %>%
  group_by(year, patch, control_treatment, camera_num, date_video, exp_phase) %>%
  summarise(video_hours = sum(video_length)) %>%
  ungroup()

#Number of hours, summarized across cameras
hours_total <- hours_per_camera_date %>%
  summarise(video_hours = sum(video_hours)) %>%
  ungroup()

hours_total
```

## Step 4: Filter data for analysis of experiment

Remove videos that do not correspond to 'pre' and 'post' dates

Remove cameras without data pre and post. For the experiment, I'm interested in the pre-to-post change over time. No data from a given time period means that I cannot look at that change!
```{r step4}
#Remove dates that do not correspond to pre and post; sometimes videos were reviewed during setup, for example.
data04 <- data03 %>%
  filter(video_priority == 1)

## Which species have NA for flowers?
check07 <- data04 %>%
  filter(is.na(flowers_camera_video)) %>%
  group_by(plant_species) %>%
  summarise(n = n()) %>%
  ungroup() %>%
  arrange(desc(n))

#Identify cameras that have data from both experimental periods
cameras_pre_and_post <- data04 %>%
  distinct(year, patch, control_treatment, exp_phase, camera_id, date_video) %>%
  group_by(year, patch, control_treatment, exp_phase, camera_id) %>%
  summarise(num_dates = n()) %>%
  ungroup() %>%
  pivot_wider(names_from = exp_phase, values_from = num_dates, values_fill = 0) %>%
  filter(post > 0 & pre > 0)

data_for_experiment <- data04 %>%
  filter(camera_id %in% cameras_pre_and_post$camera_id)
```

## Step 5: Filter data for 'normal' (natural/unmanipulated) visitation

Remove data from 'post' period of treatment replicates (i.e., data affected by experimental *Heliconia* removal)
```{r step5}
data_for_metanetwork <- data03 %>%
  filter(!(control_treatment == "treatment" & exp_phase == "post"))
```

## Step 6: Filter data for analysis of sampling method

Want pollen and camera networks from the same time period, 'pre' only. Easiest just to use experiment data, subset to 'pre' period (will build pre-post interaction networks based on the experiment dataset)
```{r step6}
data_for_sampling_method <- data_for_experiment %>%
  filter(exp_phase == "pre")
```

## Step 7: Summarize data

To build interaction networks, I first need an interaction rate between each hummingbird species and plant species. Here, the interaction rate will be the sighting rate, i.e., number of hummingbird sightings per hour. A sighting occurs when a hummingbird is in the camera frame. However, it's possible that a hummingbird could just be flying through, without any intention of visiting the plant. So, I will subset sightings to those in which the bird seems to drink from ("visit") the flower. For this reason, I also call refer to the "sighting rate" as the "visitation rate".

To calculate the rate, I will total the number of hummingbirds seen and then divide by the observation effort (number of video hours).

At this point it's also good to think ahead about the different datasets I'll need:

1. To understand how hummingbird visitation changes as result of our experimental manipulation (Heliconia removal), I'll need to summarize data at the level of EXPERIMENTAL PERIOD (pre vs. post) and REPLICATE (i.e., site + year combination). To explore how sampling method (camera observations vs. pollen samples) influences network metrics from individual networks, I can use this same dataset, but filtered to the 'pre' period only.

2. To visualize 'normal' interactions within the study system, I'll need to just look at unmanipulated data from the 'pre' period. Here, I am interested in the interactions across sites and years. This is the "meta-network" (network of networks) approach.

Get files
```{r step7a}
## Source the summarizing function
source("../../code/helper_functions/Summarize_camera_data.R")

## Get scientific names for birds and plants (not just abbreviations). Will add after summarizing 
bird_names <- read.csv("../../data/import/data/Hummingbird_codes_species_list.csv")

## Plant names. Will use this for camera data only; for pollen data, do not always know exact plant species associated with each pollen morphotype
plant_names <- read.csv("../../data/import/data/Plant_codes_species_list.csv") %>%
  select(plant_species, plant_scientific_name) %>%
  filter(!is.na(plant_species) & plant_species != "") #Get rid of blank rows
```

Summarize data
```{r step7b}
## Pre-post networks
sighting_rates_for_pp_networks = data_for_experiment %>%
  calculate_sighting_rates(data = ., level_org = "plant_species", level_time = "exp_phase", level_bird = "camera_spp_separate", sightings = "with_visit", include_unknown_spp = FALSE) %>%
  left_join(bird_names) %>%
  left_join(plant_names) %>%
  select(-flowers, -subset) %>%
  select(year, patch, control_treatment, exp_phase, bird_species, plant_species, everything()) %>%
  filter(sightings > 0) #Remove empty rows

## For meta-network, first filter to data from 'pre' period, which are not affected by experiment. Could probably also include data 'post' period of control sites, but just looking at pre period for simplicity
sighting_rates_for_metanetwork = data_for_metanetwork %>%
  calculate_sighting_rates(data = ., level_org = "plant_species_across_sites", level_time = "all", level_bird = "camera_spp_separate", sightings = "with_visit", include_unknown_spp = FALSE) %>%
  left_join(bird_names) %>%
  left_join(plant_names) %>%
  select(-flowers, -subset) %>%
  filter(sightings > 0) #Remove empty rows
```

## Step 8: Summarize visitors to *Heliconia* (for general results)

What percentage of sightings are from green hermits and violet sabrewings?
```{r step8}
#Include unknown species in this summary
sighting_rates_heto <- data_for_metanetwork %>%
  calculate_sighting_rates(data = ., level_org = "plant_species_across_sites", level_time = "all", level_bird = "camera_spp_separate", sightings = "with_visit", include_unknown_spp = TRUE) %>%
  left_join(bird_names) %>%
  left_join(plant_names) %>%
  filter(plant_species == "HETO")

num_sightings_heto <- sighting_rates_heto %>%
  summarise(num_sightings = sum(sightings))

num_sightings_heto_greh_visa <- sighting_rates_heto %>%
  filter(bird_species == "GREH" | bird_species == "VISA") %>%
  summarise(num_sightings = sum(sightings))

#Calculate % sightings from GREH/VISA
num_sightings_heto
num_sightings_heto_greh_visa
num_sightings_heto_greh_visa/num_sightings_heto
```

## Step 9: Summarize frequency of robbing (for general results)

How many sightings involve robbing? Summarized for experiment
```{r step9}
honest_vs_rob <- data_for_experiment %>%
  filter(exp_phase == "pre") %>%
  filter(sightings_yes_or_no == "Y") %>%
  filter(visit_type != "none") %>%
  filter(visit_type != "unknown_type") %>%
  group_by(visit_type) %>%
  summarise(num_visits = n()) %>%
  ungroup()
  
honest_vs_rob

#Percentage of visits that involve robbing (out of total where visit type could be determined)
(total_confirmed = sum(honest_vs_rob$num_visits))
(total_rob = honest_vs_rob %>% filter(visit_type == "rob") %>% pull(num_visits))
(total_honest_rob = honest_vs_rob %>% filter(visit_type == "honest_and_rob") %>% pull(num_visits))
(total_rob + total_honest_rob)/total_confirmed
```

## Step 10: Export
```{r step10}
## Camera data, filtered but not summarized
write.csv(data_for_experiment, "../../data/export/intermediate/Camera_data_filtered_for_experiment.csv", row.names = FALSE)
write.csv(data_for_metanetwork, "../../data/export/intermediate/Camera_data_filtered_for_metanetwork.csv", row.names = FALSE) #'normal' visitation
write.csv(data_for_sampling_method, "../../data/export/intermediate/Camera_data_filtered_for_sampling_method.csv", row.names = FALSE)

## Camera data, summarized for analyses using pre and post networks from individual replicates
write.csv(sighting_rates_for_pp_networks, "../../data/export/for_analysis/Camera_data_summarized_for_pp_networks.csv", row.names = FALSE)

## Camera data, summarized for visualization of meta-network
write.csv(sighting_rates_for_metanetwork, "../../data/export/for_analysis/Camera_data_summarized_for_metanetwork.csv", row.names = FALSE) 
```
